python-whisper/requirements.txt
openai-whisper
flask
flask-cors
torch
torchaudio